break up utils into two files
also rename _check_fields -> check_fields
