+++ propagator/utils/gis.py
@@ -1,7 +1,7 @@
-""" Basic utility functions for ``propagator``.
+""" Wrappers around Esri's ``arcpy`` library.
 
 This contains basic file I/O, coversion, and spatial analysis functions
-to support the python-tidegates library. In most cases, these functions
+to support the python-propagator library. These functions generally
 are simply wrappers around their ``arcpy`` counter parts. This was done
 so that in the future, these functions could be replaced with calls to
 a different geoprocessing library and eventually ween the code base off
@@ -17,14 +17,14 @@ Written by Paul Hobson (phobson@geosyntec.com)
 
 
 import os
-import datetime
 import itertools
-from functools import wraps
 from contextlib import contextmanager
 
 import numpy
 
 import arcpy
+
+from .misc import update_status
 
 
 class RasterTemplate(object):
@@ -307,37 +307,6 @@ def WorkSpace(path):
     arcpy.env.workspace = orig_workspace
 
 
-def _status(msg, verbose=False, asMessage=False, addTab=False): # pragma: no cover
-    if verbose:
-        if addTab:
-            msg = '\t' + msg
-        if asMessage:
-            arcpy.AddMessage(msg)
-        else:
-            print(msg)
-
-
-def update_status(): # pragma: no cover
-    """ Decorator to allow a function to take a additional keyword
-    arguments related to printing status messages to stdin or as arcpy
-    messages.
-
-    """
-
-    def decorate(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            msg = kwargs.pop("msg", None)
-            verbose = kwargs.pop("verbose", False)
-            asMessage = kwargs.pop("asMessage", False)
-            addTab = kwargs.pop("addTab", False)
-            _status(msg, verbose=verbose, asMessage=asMessage, addTab=addTab)
-
-            return func(*args, **kwargs)
-        return wrapper
-    return decorate
-
-
 def create_temp_filename(filepath, filetype=None, prefix='_temp_', num=None):
     """ Helper function to create temporary filenames before to be saved
     before the final output has been generated.
@@ -393,7 +362,7 @@ def create_temp_filename(filepath, filetype=None, prefix='_temp_', num=None):
     return os.path.join(ws, folder, prefix + filename + num + ext)
 
 
-def _check_fields(table, *fieldnames, **kwargs):
+def check_fields(table, *fieldnames, **kwargs):
     """
     Checks that field are (or are not) in a table. The check fails, a
     ``ValueError`` is raised. Relies on `arcpy.ListFields`_.
@@ -840,43 +809,6 @@ def aggregate_polygons(polygons, ID_field, filename):
     return dissolved
 
 
-@update_status() # array
-def flood_zones(zones_array, topo_array, elevation):
-    """ Mask out non-flooded portions of arrays.
-
-    Parameters
-    ----------
-    zones_array : numpy.array
-        Array of zone IDs from each zone of influence.
-    topo_array : numpy.array
-        Digital elevation model (as an array) of the areas.
-    elevation : float
-        The flood elevation *above* which everything will be masked.
-
-    Returns
-    -------
-    flooded_array : numpy.array
-        Array of zone IDs only where there is flooding.
-
-    """
-
-    # compute mask of non-zoned areas of topo
-    nonzone_mask = zones_array <= 0
-
-    invalid_mask = numpy.ma.masked_invalid(topo_array).mask
-    topo_array[invalid_mask] = -999
-
-    # mask out zoned areas above the flood elevation
-    unflooded_mask = topo_array > elevation
-
-    # apply the mask to the zone array
-    final_mask = nonzone_mask | unflooded_mask
-    flooded_array = zones_array.copy()
-    flooded_array[final_mask] = 0
-
-    return flooded_array
-
-
 @update_status() # None
 def add_field_with_value(table, field_name, field_value=None,
                          overwrite=False, **field_opts):
@@ -934,7 +866,7 @@ def add_field_with_value(table, field_name, field_value=None,
     field_type = field_opts.pop("field_type", typemap[type(field_value)])
 
     if not overwrite:
-        _check_fields(table, field_name, should_exist=False)
+        check_fields(table, field_name, should_exist=False)
 
     if field_value is None and field_type is None:
         raise ValueError("must provide a `field_type` if not providing a value.")
@@ -1040,6 +972,10 @@ def load_attribute_table(input_path, *fields):
     """
     Loads a shapefile's attribute table as a numpy record array.
 
+    Relies on `arcpy.da.TableToNumPyArray`_.
+
+    .. _arcpy.da.TableToNumPyArray: http://goo.gl/NzS6sB
+
     Parameters
     ----------
     input_path : str
@@ -1053,6 +989,10 @@ def load_attribute_table(input_path, *fields):
     -------
     records : numpy.recarray
         A record array of the selected fields in the attribute table.
+
+    See also
+    --------
+    groupby_and_aggregate
 
     Examples
     --------
@@ -1073,7 +1013,7 @@ def load_attribute_table(input_path, *fields):
     layer = load_data(input_path, "layer")
 
     # check that fields are valid
-    _check_fields(layer.dataSource, *fields, should_exist=True)
+    check_fields(layer.dataSource, *fields, should_exist=True)
 
     array = arcpy.da.FeatureClassToNumPyArray(in_table=input_path, field_names=fields)
     return array
@@ -1086,11 +1026,6 @@ def groupby_and_aggregate(input_path, groupfield, valuefield,
     Counts the number of distinct values of `valuefield` are associated
     with each value of `groupfield` in a data source found at
     `input_path`.
-
-    Relies on `arcpy.da.TableToNumPyArray`_.
-
-    .. _arcpy.da.TableToNumPyArray: http://goo.gl/NzS6sB
-
 
     Parameters
     ----------
@@ -1135,6 +1070,7 @@ def groupby_and_aggregate(input_path, groupfield, valuefield,
     --------
     itertools.groupby
     populate_field
+    load_attribute_table
 
     """
 
@@ -1220,7 +1156,7 @@ def populate_field(table, value_fxn, valuefield, *keyfields):
 
     fields = list(keyfields)
     fields.append(valuefield)
-    _check_fields(table, *fields, should_exist=True)
+    check_fields(table, *fields, should_exist=True)
 
     with arcpy.da.UpdateCursor(table, fields) as cur:
         for row in cur:
@@ -1270,7 +1206,7 @@ def copy_data(destfolder, *source_layers, **kwargs):
     return copied
 
 
-@update_status()
+@update_status() # layer
 def concat_results(destination, *input_files):
     """ Concatentates (merges) serveral datasets into a single shapefile
     or feature class.
@@ -1300,7 +1236,7 @@ def concat_results(destination, *input_files):
     return result_to_layer(result)
 
 
-@update_status()
+@update_status() # layer
 def join_results_to_baseline(destination, result_file, baseline_file):
     """ Joins attributes of a geoprocessing result to a baseline dataset
     and saves the results to another file.
@@ -1339,3 +1275,5 @@ def join_results_to_baseline(destination, result_file, baseline_file):
     )
 
     return result_to_layer(result)
+
+
